*moved char buffer to display file *added comments for global pid variables *improved parameter menu loop to support changing drop off pwm

+++ src/tapefollowing.cpp
@@ -32,7 +32,7 @@ const float turning_multiplier = 40;
  
 // Speed-dependent variables;
 int max_pwm = 950;
-float multiplier = 25;
+int multiplier = 25;
  
 // PID variables:
 volatile int lastErrState = 0;
